+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Keyword `:flat` to skip setting up completion and syntax checking.
+- The defaults for `:completion` are now customizable through
+  `harpoon-completion-{auto,delay,prefix}`.
 
 ### Removed